requirements.txt
flask
numpy
tensorflow
Pillow
keras
sklearn

